update archivemount wrapper

+++ cda.sh
@@ -1,15 +1,20 @@
-#!/bin/sh
+#!/bin/bash
+# archivemount wrapper which implements cda
+
 export TEMPDIR=`mktemp -d`
 export SHELL="bash"
-export UID=`id -u`
-export GID=`id -g`
 
 set -e
 
-archivemount -o ro -o uid=${UID} -o gid=${GID} -o nonempty $@ ${TEMPDIR} 
-cd ${TEMPDIR}
-export TEMPDIR=`pwd`
-${SHELL}
-cd /
-umount ${TEMPDIR}
-rmdir ${TEMPDIR}
+clean () {
+	if [ -d $TEMPDIR ]; then 
+		if ! $(mountpoint ${TEMPDIR}) ; then
+			fusermount -u ${TEMPDIR};
+		fi
+		rmdir ${TEMPDIR};
+	fi
+}
+trap clean EXIT	
+
+archivemount -o ro -o nonempty $@ ${TEMPDIR} 
+(cd ${TEMPDIR}; ${SHELL})
